+++ crates/analyzer/Cargo.toml
@@ -9,8 +9,6 @@ edition = "2018"
 path = "src/lib.rs"
 
 [dependencies]
-data-encoding = "2.3.1"
-ring = "0.16.19"
 nom = "6.1.0"
 dbus = "0.9.1"
 lmdb-rkv = "0.14.0"
@@ -18,3 +16,6 @@ confy = "0.4.0"
 serde = { version = "1.0", features = ["derive"] }
 directories = "2.0.2"
 thiserror = "1.0"
+
+fapolicy-trust = { path = "../trust" }
+fapolicy-util = {path = "../util" }
