+++ crates/analyzer/src/check.rs
@@ -1,12 +1,12 @@
 use std::fs::File;
 use std::io::BufReader;
 
-use crate::api::Trust;
 use crate::app::State;
 use crate::error::Error;
 use crate::error::Error::{FileNotFound, MetaError};
-use crate::sha::sha256_digest;
-use crate::trust::{Actual, Status};
+use fapolicy_trust::stat::{Actual, Status};
+use fapolicy_trust::trust::Trust;
+use fapolicy_util::sha::sha256_digest;
 use std::process::Command;
 use std::time::UNIX_EPOCH;
