+++ crates/analyzer/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod api;
 pub mod app;
 pub mod cfg;
 pub mod check;
@@ -7,8 +6,6 @@ pub mod fapolicyd;
 pub mod log;
 pub mod rpm;
 pub mod rules;
-pub mod sha;
 pub mod svc;
 pub mod sys;
-pub mod trust;
 pub mod users;
