+++ crates/trust/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "fapolicy-trust"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lmdb-rkv = "0.14.0"
+serde = { version = "1.0", features = ["derive"] }
+thiserror = "1.0"
+
+fapolicy-util = { path = "../util" }
