+++ crates/trust/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod error;
+pub mod ops;
+pub mod read;
+pub mod stat;
+
+pub mod trust;
