+++ crates/trust/src/stat.rs
@@ -1,0 +1,19 @@
+use crate::trust::Trust;
+
+/// Actual delivers metadata about the actual file that exists on the filesystem.
+/// This is used to identify discrepancies between the trusted and the actual files.
+#[derive(Clone)]
+pub struct Actual {
+    pub size: u64,
+    pub hash: String,
+    pub last_modified: u64,
+}
+
+/// Trust status tag
+#[derive(Clone)]
+pub enum Status {
+    /// Filesystem matches trust
+    Trusted(Trust, Actual),
+    /// filesystem does not match trust
+    Discrepancy(Trust, Actual),
+}
