+++ crates/util/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "fapolicy-util"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+data-encoding = "2.3.1"
+ring = "0.16.19"
+thiserror = "1.0"
