+++ crates/util/src/sha.rs
