+++ tests/tests.rs
@@ -1,4 +1,3 @@
 mod log_parse;
 mod rule_file_parsing;
-mod sha256_test;
 mod trust_model_test;
